# feat: improve logging and add optional perf logging